# consumer bench same number of txs per iteration