#### Problem
Comment https://github.com/solana-labs/solana/pull/34717#discussion_r1448378961 for similar benchmark of the block-validation execution path. Bench results of different batch sizes are directly comparable if we always execute the same number of txs per bench iteration.

#### Summary of Changes
Run same number of txs per bench iteration in consumer benchmarks.

#### Sample result

before

```
running 6 tests
test bench_process_and_record_transactions_full_batch                        ... bench:     940,906 ns/iter (+/- 150,107)
test bench_process_and_record_transactions_full_batch_disable_tx_cost_update ... bench:     954,505 ns/iter (+/- 148,716)
test bench_process_and_record_transactions_half_batch                        ... bench:     486,423 ns/iter (+/- 95,842)
test bench_process_and_record_transactions_half_batch_disable_tx_cost_update ... bench:     495,185 ns/iter (+/- 117,344)
test bench_process_and_record_transactions_unbatched                         ... bench:      38,111 ns/iter (+/- 5,206)
test bench_process_and_record_transactions_unbatched_disable_tx_cost_update  ... bench:      39,969 ns/iter (+/- 11,277)

```

after

```
running 6 tests
test bench_process_and_record_transactions_full_batch                        ... bench:     906,828 ns/iter (+/- 125,829)
test bench_process_and_record_transactions_full_batch_disable_tx_cost_update ... bench:     938,470 ns/iter (+/- 181,095)
test bench_process_and_record_transactions_half_batch                        ... bench:     961,736 ns/iter (+/- 209,425)
test bench_process_and_record_transactions_half_batch_disable_tx_cost_update ... bench:     968,375 ns/iter (+/- 162,750)
test bench_process_and_record_transactions_unbatched                         ... bench:   2,492,309 ns/iter (+/- 718,294)
test bench_process_and_record_transactions_unbatched_disable_tx_cost_update  ... bench:   2,528,124 ns/iter (+/- 566,048)

```

Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->